+++ ospray/camera/Camera.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "common/Managed.h"
+#include "common/MotionTransform.h"
 #include "common/Util.h"
 
 namespace ospray {
@@ -21,14 +21,13 @@ struct OSPRAY_SDK_INTERFACE ProjectedPoint
 //! base camera class abstraction
 /*! the base class itself does not do anything useful; look into
     perspectivecamera etc for that */
-struct OSPRAY_SDK_INTERFACE Camera : public ManagedObject
+struct OSPRAY_SDK_INTERFACE Camera : public MotionTransform
 {
   Camera();
-  virtual ~Camera() override = default;
+  ~Camera() override;
 
-  virtual std::string toString() const override;
-
-  virtual void commit() override;
+  std::string toString() const override;
+  void commit() override;
 
   // Project the world space point to the screen, and return the screen-space
   // point (in normalized screen-space coordinates) along with the depth
@@ -40,9 +39,11 @@ struct OSPRAY_SDK_INTERFACE Camera : public ManagedObject
 
   // Data members //
 
-  vec3f pos; // position of the camera in world-space
-  vec3f dir; // main direction of the camera in world-space
-  vec3f up; // up direction of the camera in world-space
+  // if motionBlur in local camera space; otherwise in world-space:
+  vec3f pos; // position of the camera
+  vec3f dir; // main direction of the camera
+  vec3f up; // up direction of the camera
+
   float nearClip{1e-6f}; // near clipping distance
   // definition of the image region, may even be outside of [0..1]^2
   // to simulate sensor shift
@@ -50,10 +51,14 @@ struct OSPRAY_SDK_INTERFACE Camera : public ManagedObject
   vec2f imageEnd; // upper right corner
   range1f shutter{0.5f, 0.5f}; // start and end time of camera shutter time
 
+  void setDevice(RTCDevice embreeDevice);
+
  private:
   template <typename BASE_CLASS, typename CHILD_CLASS>
   friend void registerTypeHelper(const char *type);
   static void registerType(const char *type, FactoryFcn<Camera> f);
+  RTCDevice embreeDevice{nullptr};
+  RTCGeometry embreeGeometry{nullptr};
 };
 
 OSPTYPEFOR_SPECIALIZATION(Camera *, OSP_CAMERA);